add check for upper bounds
